+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
- - Read only `ZipStore` behind `zip` feature
+ - Read only `ZipStorageAdapter` behind `zip` feature
  - Add `StorageValueIO`, a [`std::io::Read`] interface to a storage value.
 
 ### Changed